// packages/auto-generated-code/src/common-abi-types/GmxPositionRouter.ts
export interface ExecuteDecreasePositionEventObject {
  account: string;
  path: string[];
  indexToken: string;
  collateralDelta: BigNumber;
  sizeDelta: BigNumber;
  isLong: boolean;
  receiver: string;
  acceptablePrice: BigNumber;
  minOut: BigNumber;
  executionFee: BigNumber;
  blockGap: BigNumber;
  timeGap: BigNumber;
}
export type ExecuteDecreasePositionEvent = TypedEvent<
  [
    string,
    string[],
    string,
    BigNumber,
    BigNumber,
    boolean,
    string,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ],
  ExecuteDecreasePositionEventObject
>;

export type ExecuteDecreasePositionEventFilter =
  TypedEventFilter<ExecuteDecreasePositionEvent>;

export interface ExecuteIncreasePositionEventObject {
  account: string;
  path: string[];
  indexToken: string;
  amountIn: BigNumber;
  minOut: BigNumber;
  sizeDelta: BigNumber;
  isLong: boolean;
  acceptablePrice: BigNumber;
  executionFee: BigNumber;
  blockGap: BigNumber;
  timeGap: BigNumber;
}
export type ExecuteIncreasePositionEvent = TypedEvent<
  [
    string,
    string[],
    string,
    BigNumber,
    BigNumber,
    BigNumber,
    boolean,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ],
  ExecuteIncreasePositionEventObject
>;